# apps/api/apiSwagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/orders":{"post":{"operationId":"OrdersController_create","parameters":[],"requestBody":{"required":true,"description":"Create order","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/CreateLimitOrder"},{"$ref":"#/components/schemas/CreateMarketOrder"}]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}},"403":{"description":""}}},"get":{"operationId":"OrdersController_getOrders","parameters":[{"name":"page","required":false,"in":"query","description":"Page to be query","schema":{"type":"number"}},{"name":"pageSize","required":false,"in":"query","description":"Page size to be query","schema":{"type":"number"}},{"name":"status","required":false,"in":"query","description":"Status to be included, by default get all status","schema":{"type":"array","items":{"type":"string"}}},{"name":"side","required":false,"in":"query","description":"Order side to be included","schema":{"type":"string"}},{"name":"externalId","required":false,"in":"query","description":"External ID to be filter","schema":{"type":"number"}},{"name":"market","required":false,"in":"query","description":"The market pair to be query","schema":{"type":"string"}},{"name":"orderType","required":false,"in":"query","description":"The order type to filter","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}},"403":{"description":""}}}},"/orders/cancel/{id}":{"put":{"operationId":"OrdersController_cancel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}},"403":{"description":""}}}},"/orders/books/{market}":{"get":{"operationId":"OrdersController_getBook","parameters":[{"name":"market","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Books"}}}},"403":{"description":""}}}},"/tenants":{"post":{"operationId":"TenantsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTenantDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantDto"}}}},"403":{"description":""}}}},"/tenants/engine":{"put":{"operationId":"TenantsController_createEngine","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddEngineDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantDto"}}}},"403":{"description":""}}}},"/tenants/webhook":{"post":{"operationId":"TenantsController_addWebhook","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddWebhookDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebhookDto"}}}},"403":{"description":""}}}},"/trades":{"get":{"operationId":"TradesController_getTrades","parameters":[{"name":"page","required":false,"in":"query","description":"Page to be query","schema":{"type":"number"}},{"name":"pageSize","required":false,"in":"query","description":"Page size to be query","schema":{"type":"number"}},{"name":"market","required":false,"in":"query","description":"The market pair to be query","schema":{"type":"string"}},{"name":"createdAt","required":false,"in":"query","description":"The oldest date-time to be filter","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradeDto"}}}},"403":{"description":""}}}},"/analytics":{"get":{"operationId":"AnalyticsController_getAnalytics","parameters":[{"name":"market","required":true,"in":"query","description":"The market to be queried","schema":{"type":"string"}},{"name":"startTime","required":false,"in":"query","description":"The start time of the search","schema":{"format":"date-time","type":"string"}},{"name":"endTime","required":false,"in":"query","description":"The start time of the search","schema":{"format":"date-time","type":"string"}},{"name":"limit","required":false,"in":"query","description":"The start time of the search","schema":{"maximum":500,"default":100,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KLineDto"}}}},"403":{"description":""}}}}},"info":{"title":"Matching engine document","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateLimitOrder":{"type":"object","properties":{"market":{"type":"string","description":"The market pair of the order"},"side":{"type":"string","description":"Side of the order","enum":["BID","ASK"]},"orderType":{"type":"string","description":"Type of the order","enum":["LIMIT","MARKET"]},"volume":{"type":"number","description":"Volume that should be traded"},"price":{"type":"number","description":"Price that should be traded"},"externalId":{"type":"number","description":"Client side ID"}},"required":["market","side","orderType","volume","price","externalId"]},"CreateMarketOrder":{"type":"object","properties":{"market":{"type":"string","description":"The market pair of the order"},"side":{"type":"string","description":"Side of the order","enum":["BID","ASK"]},"orderType":{"type":"string","description":"Type of the order","enum":["LIMIT","MARKET"]},"volume":{"type":"number","description":"Volume that should be traded"},"externalId":{"type":"number","description":"Client side ID"}},"required":["market","side","orderType","volume","externalId"]},"OrderDto":{"type":"object","properties":{"market":{"type":"string","description":"The market pair of the order"},"side":{"type":"string","description":"Side of the order","enum":["BID","ASK"]},"orderType":{"type":"string","description":"Type of the order","enum":["LIMIT","MARKET"]},"volume":{"type":"number","description":"Volume that should be traded"},"price":{"type":"number","description":"Price that should be traded"},"externalId":{"type":"number","description":"Client side ID"},"id":{"type":"number","description":"Unique identifier of a single kline"},"locked":{"type":"number","description":"The expected value to be locked from original account"},"status":{"type":"string","description":"Status of the Order","enum":["OPENED","FILLED","CANCELED","PARTIALLY_FILLED"]},"tradedVolume":{"type":"number","description":"How much of volume was traded"},"tradesCount":{"type":"number","description":"In how many trades this order appears"},"tenantId":{"type":"number","description":"The tenant who originate it"},"createdAt":{"format":"date-time","type":"string","description":"The time in which this order was created"}},"required":["market","side","orderType","volume","price","externalId","id","locked","status","tradedVolume","tradesCount","tenantId","createdAt"]},"Books":{"type":"object","properties":{"ask":{"description":"The books with the accumulated ask orders","type":"array","items":{"type":"string"}},"bid":{"description":"The books with the accumulated bid orders","type":"array","items":{"type":"string"}}},"required":["ask","bid"]},"CreateTenantDto":{"type":"object","properties":{"availableMarkets":{"type":"string","description":"Markets available on your matching engine, use the format `btcbrl, ethbrl`."}},"required":["availableMarkets"]},"TenantDto":{"type":"object","properties":{"availableMarkets":{"type":"string","description":"Markets available on your matching engine, use the format `btcbrl, ethbrl`."},"id":{"type":"number","description":"Unique identifier of the tenant"},"accessUuid":{"type":"string","description":"Uuid to be used on all requests"}},"required":["availableMarkets","id","accessUuid"]},"AddEngineDto":{"type":"object","properties":{"newMarkets":{"type":"string","description":"The new engines that will be added to the system"}},"required":["newMarkets"]},"AddWebhookDto":{"type":"object","properties":{"fullAddress":{"type":"string","description":"The address to where the webhook will be send."},"action":{"type":"string","description":"The possible actions to invoke the webhook","enum":["BOOKUPDATE","NEWTRADES"]}},"required":["fullAddress","action"]},"WebhookDto":{"type":"object","properties":{"fullAddress":{"type":"string","description":"The address to where the webhook will be send."},"action":{"type":"string","description":"The possible actions to invoke the webhook","enum":["BOOKUPDATE","NEWTRADES"]},"id":{"type":"number","description":"Unique identifier of the tenant"},"tenantId":{"type":"number","description":"The tenant who originate it"}},"required":["fullAddress","action","id","tenantId"]},"TradeDto":{"type":"object","properties":{"id":{"type":"number","description":"Unique identifier of a single kline"},"market":{"type":"string","description":"The market pair of the trade"},"price":{"type":"number","description":"Price traded"},"volume":{"type":"number","description":"Volume traded"},"funds":{"type":"number","description":"Value used on original account"},"makerId":{"type":"number","description":"Id of order who was in the book."},"takerId":{"type":"number","description":"Id of order was beeing process."},"tenantId":{"type":"number","description":"The tenant who originate it"},"createdAt":{"format":"date-time","type":"string","description":"The time in which this order was created"}},"required":["id","market","price","volume","funds","makerId","takerId","tenantId","createdAt"]},"KLineDto":{"type":"object","properties":{"id":{"type":"number","description":"Unique identifier of a single kline"},"tenantId":{"type":"number","description":"The tenant who originate it"},"market":{"type":"string","description":"The market to which refers to"},"openTime":{"format":"date-time","type":"string","description":"The time of the first trade registered in this kline"},"openValue":{"type":"number","description":"The value of the first trade on the kline"},"highestValue":{"type":"number","description":"The highest value hitted between the alive time fot his kline"},"lowestValue":{"type":"number","description":"The lowest value hitted between the alive time fot his kline"},"closeValue":{"type":"number","description":"The value of the last trade on the kline"},"closeTime":{"format":"date-time","type":"string","description":"The time of the last trade registered in this kline"},"numberTrades":{"type":"number","description":"The number of trades tha happened between the alive time fot his kline"},"tradedVolume":{"type":"number","description":"The volume traded between the alive time fot his kline"},"lastTradedSecond":{"type":"number","description":"The close time in seconds, always rounded for multiples of 5(because a new kline is created every 5 seconds, when have a new trade)"}},"required":["id","tenantId","market","openTime","openValue","highestValue","lowestValue","closeValue","closeTime","numberTrades","tradedVolume","lastTradedSecond"]}}}}