# apps/api/src/tenants/dto/add-engine.dto.ts
import { ApiProperty } from "@nestjs/swagger";

export class AddEngineDto {
  @ApiProperty({
    description: 'The new engines that will be added to the system',
    required: true
  })
  newMarkets: string;
}